pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = ""
authors = ["codcod <nikos@linux.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sanic = "^23.6.0"
sanic-ext = "^23.6.0"
opentelemetry-api = "^1.20.0"
opentelemetry-sdk = "^1.20.0"
opentelemetry-propagator-b3 = "^1.20.0"
opentelemetry-exporter-prometheus = "^1.12.0rc1"

[tool.poetry.group.dev.dependencies]
pytest-coverage = "^0.0"
black = "^23.9.1"
flake8 = "^6.1.0"
isort = "^5.12.0"
pylint = "^3.0.1"

[tool.black]
line-length = 88
skip-string-normalization = 1
target-version = ['py310']

[tool.pylint.basic]
# Allow shorter and longer variable names than the default.
argument-rgx = "[a-z_][a-z0-9_]*$"
attr-rgx = "[a-z_][a-z0-9_]*$"
variable-rgx = "[a-z_][a-z0-9_]*$"

[tool.pylint.messages_control]
disable = [
    "empty-docstring",
    "missing-class-docstring",
    "missing-function-docstring",
    "missing-module-docstring"
]
ignore = ["setup.py", "__init__.py"]
ignore-paths = ['.*/tests', ]

[tool.pylint.format]
max-line-length = 88

[tool.isort]
profile = "black"
line_length = 88
auto_identify_namespace_packages = false
force_single_line = true # pyright doesn't like implicit namespace + single line (related to https://github.com/microsoft/pyright/issues/2882?)
known_first_party = ["bookstore"]

[tool.pytest.ini_options]
# https://docs.pytest.org/en/latest/reference/customize.html
minversion = "6.0"
addopts = "--doctest-modules -vv --cov=bookstore tests"
testpaths = [
    "tests",
]
